Don't elaborate terms eagerly in Rewrites.explicit
This can cause problems when elaborating global constant expecting type class instances as Lean errors out when it fails to synthesize such an instance during elaboration. To avoid this, we lookup the value and type of global constants in the environment. We then only need to be carful to instantiate level parameters with fresh level mvars (this seems to happen automatically when explicitly elaborating such a constant).